Split MofGeneratorTests and ParserTests into multiple partial class files
