+++ src/Kingsland.MofParser.UnitTests/CodeGen/RoundtripTests_BooleanValue.cs
@@ -1,0 +1,31 @@
+﻿using NUnit.Framework;
+
+namespace Kingsland.MofParser.UnitTests.CodeGen
+{
+
+    public static partial class RoundtripTests
+    {
+
+        #region 7.6.1.5 Boolean value
+
+        public static class BooleanValueTests
+        {
+
+            [Test]
+            public static void BooleanValueAstShouldRoundtrip()
+            {
+                RoundtripTests.AssertRoundtrip(
+                    "instance of myType as $Alias00000070\r\n" +
+                    "{\r\n" +
+                    "\tReference = TRUE;\r\n" +
+                    "};"
+                );
+            }
+
+        }
+
+        #endregion
+
+    }
+
+}
